Updated build process
This patch updates the build process so that the protobuf generation is
handled via a Makefile instead of directly in the Travis-CI build
script. This makes it easier to generate the protobuf locally. The
Makefile includes "csi.mk", which contains the generation logic and can
be reused by other projects.

Additionally the Travis-CI build script was updated to explicitly
declare its expected package location within Go's package structure.
This is useful for forks of this repository that set up their own
Travis-CI builds.

Finally the Travis-CI build script has been updated to push the
generated protobuf and associated Go source file to a Gist in order to
make them available after a build.