Version the CSI protobuf
This patch versions the generated CSI protobuf and provides a Travis-CI
build step that verifies the protobuf is up-to-date before allowing any
changes to be merged.

Additionally this patch also adds minimal Go language bindings in order
to verify the specification. Finally, scaffolding was put into place in
order to support Cxx language bindings in the future as well.