Remove Lang Binding Builds from Root Makefile
This patch removes part of the root Makefile that would call the
language-binding builds when running `make` locally from the root of the
project. If the `lib/go` build was executed outside of the GOPATH errors
would occur. Since the repository is not Go-specific, running a
Go-specific build by default should not be assumed.