spec: Move `AccessMode` to `VolumeCapability`
This patch moved `AccessMode` to `VolumeCapability`, because access mode
is strongly related to the filesystem type. For instance, a plugin might
support `SINGLE_NODE_WRITER` for `xfs`, and `MULTI_NODE_READER_ONLY` for
`gfs2`.

This patch also removed `VolumeCapability` fields from
`ControllerGetCapabilities` and `NodeGetCapabilities` RPCs. It becomes
very tedious to ask the plugins to enumerate all volume capabilities
that a plugin supports, and probably provides less value. To check the
capablities of a volume, the CO should use `ValidateVolumeCapabilities`.
This applies to both newly provisioned volumes and pre-existing volumes.

See discussion in #61.