spec: Add volume capabilities and parameters to GetCapacityRequest
This change allows the CO to get the capacity of a storage pool with the
given volume capabilities and parameters. This is particularly useful
when the CO defines a storage class or profile, and wants to know the
capacity of the storage under that class or profile so that it can
perform allocation or track utilization.

For example, assume that a Plugin can provision both raid0 and raid1
volumes. Depending on the type, the capacity is different. This patch
allows the CO to query the capacity of the storage pool for each type:

```
request:
  version:
    major: 0
    minor: 1
    patch: 0
  volume_capabilities:
    - mount:
      fs_type: xfs
  parameters:
    type: raid1
response:
  result:
    total_capacity: 100000
```

```
request:
  version:
    major: 0
    minor: 1
    patch: 0
  volume_capabilities:
    - mount:
      fs_type: xfs
  parameters:
    type: raid0
response:
  result:
    total_capacity: 200000
```