spec: Fix the error codes related to idempotency
The patch fixed the error codes that are related to idempotency:

(1) For `CreateVolume`, if the volume already exists and is compatible,
return OK instead. If the volume exists but not compatible, return
ALREADY_EXISTS.

(2) For `DeleteVolume`, if the volume does not exist, return OK instead.

(3) For `ControllerUnpublishVolume`, if the volume is already detached
from the node, return OK instead.

Fixes #157
xref #158