Remove all SP Implementation Info
This patch removes all of the SP implementation information previously
detailed in the specification. With respect to issue #171 and comments
made by @saad-ali, @childsb, @jdef, and others, it has become apparent
that the specification should not supply any suggestions or restrictions
on matters related to an SP's packaging, deployment, or runtime
configuration.

Removed information includes, but not limited to:

* `CSI_ENDPOINT` - SPs are now responsible for defining the
  manner by which they grok the gRPC endpoint on which they serve
  the CSI services. CO admins are now responsible for deploying
  and configuring SPs using that SP's documentation.

* Env var names - The spec no longer claims ownership of the `CSI_`
  prefix for environment variables.

* Permissions - The spec no longer asserts any recommendations with
  respect to container permissions that an SP might require.

* Logging - All logging details are the responsibility of the SP and
  may vary SP to SP.

The removed information may very well be suitable to live in a separate
document that provides guidelines for SP creation and configuration,
such as common environment variable names, logging and packaging details,
etc.

cc @codenrhoden