spec: Fix SnapshotSource commment
Comment in SnapshotSource make references as if it were a newly created
snapshot instead of a reference to a source snapshot.

Talking about "created snapshot" and how this id will be used in
subsequent calls may be confusing.

Comment is changed to reflect that it is an already existing source
snapshot.