spec: Volume from snapshot with different size
There's no mention in the spec to expected behavior when the
required_bytes in CreateVolumeRequest's capacity_range is larger or
smaller than that of the VolumeContentSource.

This may lead to different interpretations on the behavior.

We are now explicit about expected behavior. If it's smaller we fail
with OUT_OF_RANGE, and if it's greater the volume must be extended by
the driver if necessary.