spec: Add efficient volume cloning support
With the new CSI snapshot feature it's now possible to create volumes
from a snapshots.  This may lead to users abusing it for unintended
functionality, such as having a golden image to source new volumes.

Most storage backends support efficient volume cloning, so it would be
benefitial to expose this functionality in the CSI spec.

By supporting cloning we avoid users creating a volume, and then a
snapshot of that volume, just so it can be used as the source for new
volumes.

The VolumeContentSource, that was added by the snapshots feature, will
be used for this feature.  New VolumeSource message is added to the
VolumeContentSource.