doc: Add comment about scalars in proto3
GRPC proto3 docs states that all scalar fields have a default value, and
that default values for scalars will not be transferred over the wire.

This makes it impossible to know if a message had a field set in the
first place, so even if a field is REQUIRED in the spec and it is set on
send, it may not be present on the message.

This is something that developers must be aware when writing plugins, as
they should not be check for the presence of scalar REQUIRED fields on
the message, as they'll only be present when the value is not the
default.

In this patch we add a comment at the beginning of the document for
those who are not aware of this implementation detail of proto3.