spec: Clarify target_path and staging_target_path
This patch clarifies the `target_path` and `staging_target_path` in the
spec. It is CO's responsibility to make sure that SP has the permission
to read or write `target_path` and `staging_target_path`, and is able to
create files/directories if the path does not exist.

This patch also changes the wording in `NodeStageVolume` call to enable
those plugins that want to use the `NodeStageVolume` call as a lifecycle
hook, but does not want to mount to the `staging_target_path`.