Update the requirements for staging and target paths
The vague wording around the staging and target paths left open too
many possible implementations for raw block volumes, and implementors
ended up doing non-obvious things. This change clarifies the contract
between the CO and the SP when staging and publishing raw block volumes
on the node.

Specifically we now require that the CO creates the staging_target_path
and that it is always a directory. And we require that the SP create
the target_path and it should be a device file for blocks and a mount
point for filesystems.