Clarify content of cloned / restored volumes
When cloning a volume or restoring a snapshot, the resulting volume should be
exact copy of the original volume. This includes any filesystem metadata.
Therefore, when the new volume is larger than the original one, the new
volume may need to resize the filesystem (ie call resize2fs, xfs_growfs, ...) to
make the bigger capacity available to workloads.