GetCapacityResponse: add maximum_volume_size
This new field has a more precisely defined semantic than the existing
available_capacity. It is intended for checking in advance in which
topology a CreateVolume call might succeed. Checking that based
on available_capacity is less precise, for example because
fragmentation might prevent creating a single volume that uses
all available capacity.

The new value has to be returned in a message to allow the caller to
determine whether the value was provided.