Change required capability for published_node_ids in GetVolume
Instead of PUBLISH_UNPUBLISH_VOLUME, use the LIST_VOLUMES_PUBLISHED_NODES
capability to control the published_node_ids field in ControllerGetVolume.
This change is to be more consistent with the existing behavior. It is
a backwards-incompatible change, but the API is alpha.

Also clarify the LIST_VOLUMES_PUBLISHED_NODES docs.