GetCapacityResponse: promote maximum_volume_size
Kubernetes and several CSI drivers have been using that field for a while now
and no changes to the semantic are expected, therefore the alpha status can and
should be removed.

The minimum_volume_size field is not used by Kubernetes. The semantic is
probably stable, but further practical experience with it might be desirable
before promoting it.