add secrets field support for ControllerGetVolume Request
At present secrets are not part of this request and some
consumers need this field for its operation.: That said,
storage backends like Ceph Cluster require the access to be
achieved with the help of authentication. For such systems this
becomes a requirement to have secrets/credentials be part of
the request CO make for its further operations or fetching the state
from the backend cluster.

Signed-off-by: Humble Chirammal <hchiramm@redhat.com>