Regenerate with protoc-gen-go 1.3.5
go.mod has github.com/golang/protobuf 1.5.3, which has a protoc-gen-go that is
not compatible with csi.pb.go in this repo.

Download github.com/golang/protobuf v1.3.5 separately, build
protoc-gen-go from it and use it to generate csi.pb.go.

The code is copied from the previous release of CSI spec, only with minor
changes.

A future update to 1.5.3 must be done carefuly and announce a breaking change.