Add note about possibly lower available capacity on MountVolume volumes
When a CreateVolume request is handled by an SP, the requested capacity
does not include any potential overhead of filesystem metadata. This can
cause confusion, as users may not be able to store data up to the amount
of the requested size of the volume.

The SP is expected to provision a volume following the capacity_range
request, but the SP may not know how much storage for filesystem
metadata will be required (depends on filesystem type and enabled
features). In these cases, the usable capacity of the volume will be
slightly smaller than the requested capacity.